Renamed magvar to magnetic declination in information dialogs. Cleanup in help menu. Now only help actions have a question mark icon. Added Qt icon. Added donation menu item in help menu and about dialog. Fixed dialog modality to keep dialog from hiding behind main window on macOS. Ctrl + cursor keys are now captured when changing details to avoid accidental zooming. Initial information dialog about missing SimConnect and databases now only shown on Windows. Fixed problems with progress dialog popping up after cancel. Added function to copy airspaces to X-Plane database. Fix for disabled connect button when using direct connect. Fix for shrinked heliports on map display. Small display fixes around weather information.

+++ src/mapgui/mappainter.h
@@ -80,6 +80,8 @@ struct PaintContext
   // Needs to be larger than number of highest level airports
   static Q_DECL_CONSTEXPR int MAX_OBJECT_COUNT = 2500;
   int objectCount = 0;
+
+  /* Increase drawn object count and return true if exceeded */
   bool objCount()
   {
     objectCount++;
@@ -124,7 +126,7 @@ struct PaintContext
 
   float szF(float scale, double size) const
   {
-    return scale * size;
+    return scale * static_cast<float>(size);
   }
 
   /* Calculate and set font based on scale */
